[add] BindingActivity, BindingFragment 추가 및 적용
