+++ app/src/main/java/org/android/go/sopt/presentation/home/HomeActivity.kt
@@ -1,20 +1,16 @@
 package org.android.go.sopt.presentation.home
 
 import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import org.android.go.sopt.R
 import org.android.go.sopt.databinding.ActivityHomeBinding
 import org.android.go.sopt.presentation.gallery.GalleryFragment
 import org.android.go.sopt.presentation.search.SearchFragment
+import org.android.go.sopt.presentation.util.binding.BindingActivity
 
-class HomeActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityHomeBinding
-
+class HomeActivity : BindingActivity<ActivityHomeBinding>(R.layout.activity_home) {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = ActivityHomeBinding.inflate(layoutInflater)
-        setContentView(binding.root)
 
         changeFragment(HomeFragment())
