+++ app/src/main/java/org/android/go/sopt/presentation/util/binding/BindingActivity.kt
@@ -1,0 +1,18 @@
+package org.android.go.sopt.presentation.util.binding
+
+import android.os.Bundle
+import androidx.annotation.LayoutRes
+import androidx.appcompat.app.AppCompatActivity
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
+
+abstract class BindingActivity<B : ViewDataBinding>(@LayoutRes private val layoutResId: Int) :
+    AppCompatActivity() {
+    lateinit var binding: B
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = DataBindingUtil.setContentView(this, layoutResId)
+    }
+
+}
