+++ app/src/main/java/org/android/go/sopt/presentation/util/binding/BindingFragment.kt
@@ -1,0 +1,32 @@
+package org.android.go.sopt.presentation.util.binding
+
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.annotation.LayoutRes
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
+import androidx.fragment.app.Fragment
+
+abstract class BindingFragment<B : ViewDataBinding>(@LayoutRes private val layoutResId: Int) :
+    Fragment() {
+    private var _binding: B? = null
+    val binding
+        get() = requireNotNull(_binding) { "${this::class.java.simpleName}에서 에러가 발생했습니다." }
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?,
+    ): View {
+        _binding = DataBindingUtil.inflate(inflater, layoutResId, container, false)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
